refactor: one-product-detail DI
